+++ Models/Customer.cs
@@ -5,7 +5,7 @@ namespace Pets.Models
 {
     public class Customer
     {
-        public Int16? ID { get; set; }
+        public int? ID { get; set; }
         public string HouseholdName { get; set; }
         public string Address { get; set; }
         public string Email { get; set; }
