+++ Models/Pet.cs
@@ -4,7 +4,7 @@ namespace Pets.Models
 {
     public class Pet
     {
-        public Int16? ID { get; set; }
+        public int? ID { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
     }
